Form and PopUp: refactoring
